Fix piwik tracking
Prevent pages from being tracked multiple times or piwik.js to be loaded
multiple times.
