+++ components/page/index.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Router from 'next/router'
+import getConfig from 'next/config'
 import NProgress from 'nprogress'
 import {translate} from 'react-i18next'
 
@@ -10,17 +11,33 @@ import Meta from '../meta'
 
 import Header from './header'
 import Footer from './footer'
-import Piwik from './piwik'
 
 Router.onRouteChangeStart = () => NProgress.start()
 Router.onRouteChangeComplete = () => NProgress.done()
 Router.onRouteChangeError = () => NProgress.done()
+
+const {publicRuntimeConfig: {
+  PIWIK_URL,
+  PIWIK_SITE_ID
+}} = getConfig()
 
 class Page extends React.PureComponent {
   static propTypes = {
     children: PropTypes.func.isRequired,
     ready: PropTypes.bool.isRequired,
     t: PropTypes.func.isRequired
+  }
+
+  componentDidMount() {
+    setTimeout(() => {
+      if (window.Piwik) {
+        const tracker = window.Piwik.getTracker(`${PIWIK_URL}/piwik.php`, PIWIK_SITE_ID)
+
+        if (tracker) {
+          tracker.trackPageView()
+        }
+      }
+    }, 300)
   }
 
   render() {
@@ -44,7 +61,6 @@ class Page extends React.PureComponent {
           </Content>
         )}
         <Footer />
-        <Piwik />
 
         <style jsx>{`
           div {
